\name{data_x}
\alias{data_x}
\docType{data}
\title{
Simulated two-dimensional regressors
}
\description{
A toy example is simulated from the uniform distribution by using \code{runif(1,0,1)}.
}
\usage{
data(data_x)
}
\format{
  A data frame with 50 observations and 2 \code{x}-variables.
}
\examples{
nrr(data_x)
}
\keyword{datasets}
