\name{data_y}
\alias{data_y}
\docType{data}
\title{
Simulated response variable
}
\description{
A toy example is simulated from the functional form of 
\code{data_y = sin(2*pi*x1)+4*(1-x2)(1+x2)+sqrt(.5)*rnorm(1), where x1 and x2 are simulated from standard uniform distribution.}
}
\usage{
data(data_y)
}
\format{
A vector of 50 observations.
}
\keyword{datasets}
