\name{data_x}
\alias{data_x}
\docType{data}
\title{
Simulated three-dimensional regressors
}
\description{
Three-dimensional regressors are simulated from the uniform distribution by using \code{runif(1,0,1)}.
}
\usage{data(data_x)}
\format{
 A data matrix with 100 observations and 3 \code{x}-variables.
}
\examples{
nrr(data_x)
}
\keyword{datasets}
