\name{data_y}
\alias{data_y}
\docType{data}
\title{
Simulated response variable
}
\description{
A toy example is simulated from the functional form of
\eqn{data_y = sin(2*pi*x1) + 4*(1-x2)*(1+x2) + sqrt(0.5) * rnorm(1)}, where \code{x1} and \code{x2} are
simulated from a uniform distribution between 0 and 1.
}
\usage{
data(data_y)
}
\format{
A vector of 50 observations
}
\keyword{datasets}
